requirements.txt
PyPDF2==3.0.1
fpdf2==2.7.8
